requirements.txt
pyspark==3.5.3
delta-spark==3.3.0
requests
pytest
apache-airflow